+++ backend/core/src/auth/services/user.service.ts
@@ -1,6 +1,7 @@
 import {
   BadRequestException,
   HttpException,
+  HttpStatus,
   Inject,
   Injectable,
   NotFoundException,
@@ -96,15 +97,19 @@ export class UserService {
     const qb = this.userRepository.getQb()
 
     if (params.filter) {
+      // eslint-disable-next-line @typescript-eslint/no-unnecessary-type-assertion
+      const user = this.req.user as User
       addFilters<Array<UserFilterParams>, typeof userFilterTypeToFieldMap>(
         params.filter,
         userFilterTypeToFieldMap,
-        distinctIDQB
+        distinctIDQB,
+        user
       )
       addFilters<Array<UserFilterParams>, typeof userFilterTypeToFieldMap>(
         params.filter,
         userFilterTypeToFieldMap,
-        qb
+        qb,
+        user
       )
     }
 
@@ -562,10 +567,29 @@ export class UserService {
   }
 
   private async authorizeUserAction(requestingUser, targetUser, action) {
+    if (requestingUser?.roles?.isJurisdictionalAdmin) {
+      return this.authorizeJurisdictionalAdmin(requestingUser, targetUser)
+    }
     return await this.authzService.canOrThrow(requestingUser, "user", action, {
       id: targetUser.id,
       jurisdictionId: targetUser.id,
     })
+  }
+
+  private authorizeJurisdictionalAdmin(requestingUser, targetUser) {
+    // jurisdictional admins can't view super admins
+    if (targetUser?.roles?.isAdmin) {
+      throw new HttpException("Forbidden", HttpStatus.FORBIDDEN)
+    }
+
+    const requesterJurisdictions = requestingUser.jurisdictions?.map((juris) => juris.id)
+    const targetJurisdictions = targetUser.jurisdictions?.map((juris) => juris.id)
+    // jurisdictional admins should only see a user if they share a jurisdiction
+    const res = requesterJurisdictions.some((juris) => targetJurisdictions.includes(juris))
+    if (!res) {
+      throw new HttpException("Forbidden", HttpStatus.FORBIDDEN)
+    }
+    return res
   }
 
   private async authorizeUserProfileAction(requestingUser, targetUser, action) {