+++ backend/core/src/units/entities/unit.entity.ts
@@ -20,13 +20,13 @@ import {
   ValidateNested,
 } from "class-validator"
 import { Expose, Type } from "class-transformer"
-import { Property } from "../../property/entities/property.entity"
 import { AmiChart } from "../../ami-charts/entities/ami-chart.entity"
 import { ValidationsGroupsEnum } from "../../shared/types/validations-groups-enum"
 import { UnitType } from "../../unit-types/entities/unit-type.entity"
 import { UnitRentType } from "../../unit-rent-types/entities/unit-rent-type.entity"
 import { UnitAccessibilityPriorityType } from "../../unit-accessbility-priority-types/entities/unit-accessibility-priority-type.entity"
 import { UnitAmiChartOverride } from "../../units/entities/unit-ami-chart-override.entity"
+import { Listing } from "../../listings/entities/listing.entity"
 
 @Entity({ name: "units" })
 class Unit {
@@ -133,11 +133,11 @@ class Unit {
   @IsString({ groups: [ValidationsGroupsEnum.default] })
   monthlyRentAsPercentOfIncome?: string | null
 
-  @ManyToOne(() => Property, (property) => property.units, {
+  @ManyToOne(() => Listing, (listing) => listing.units, {
     onDelete: "CASCADE",
     onUpdate: "CASCADE",
   })
-  property: Property
+  listing: Listing
 
   @Column({ type: "boolean", nullable: true })
   @Expose()